+++ src/elements/conditions/BlockCondition.php
@@ -19,14 +19,37 @@ class BlockCondition extends ElementCondition
      */
     protected function conditionRuleTypes(): array
     {
-        return array_merge(parent::conditionRuleTypes(), [
-            LevelConditionRule::class,
-            OwnerCategoryGroupConditionRule::class,
-            OwnerEntryTypeConditionRule::class,
-            OwnerSectionConditionRule::class,
-            OwnerTagGroupConditionRule::class,
-            OwnerUserGroupConditionRule::class,
-            OwnerVolumeConditionRule::class,
-        ]);
+        $parentConditionRuleTypes = parent::conditionRuleTypes();
+        $fieldConditionRuleTypes = [];
+
+        foreach ($parentConditionRuleTypes as $ruleType) {
+            if (!isset($ruleType['class'])) {
+                continue;
+            }
+
+            $splitClass = explode('\\', $ruleType['class']);
+            $className = __NAMESPACE__ . '\\fields\\Parent' . end($splitClass);
+
+            if (class_exists($className)) {
+                $fieldConditionRuleTypes[] = [
+                    'class' => $className,
+                    'fieldUid' => $ruleType['fieldUid'],
+                ];
+            }
+        }
+
+        return array_merge(
+            $parentConditionRuleTypes,
+            $fieldConditionRuleTypes,
+            [
+                LevelConditionRule::class,
+                OwnerCategoryGroupConditionRule::class,
+                OwnerEntryTypeConditionRule::class,
+                OwnerSectionConditionRule::class,
+                OwnerTagGroupConditionRule::class,
+                OwnerUserGroupConditionRule::class,
+                OwnerVolumeConditionRule::class,
+            ],
+        );
     }
 }